Performance: SQLDescribeCol not called unless needed (when binding None/NULL)
Performance work for Joe Salmeri, a great pyodbc user (and tester).

Unfortunately we now don't know when a column is a BIT column, so we can't convert all objects
to bools.  This was never documented and doesn't seem like a big deal since all numeric values
should still work normally.