fix: Function sequence error with prepared execute, SQLTables, same execute.  (Lost fix)
This was fixed in 2.1.12 and lost during the Python 3 conversion.

The saved prepared statement was not cleaned up when a SQLTables or similar function was used.
If the next execute used the same SQL, pyodbc did not call SQLPrepare.
(cherry picked from commit 7fdec2f4d88150cce79472c93acb9479b0ef982c)