Added context manager support to Cursor object.
Cursor objects can now be used in "with" statements and will automatically commit the
transaction if successful.

This is part of the effort to allow programs to be written using only Cursors.  There are times
when having separate connections and cursors are necessary, but very rare.  By duplicating
connection functionality in the cursor, most programs can probably use a single Cursor object
everywhere.